docker/api-data/requirements.txt
Flask
hdfs
cassandra-driver
